Language.cs: added Dutch (Nederlands)
I wish to commit a full translation in Dutch (nl_NL or nl_BE), which is Nederlands. This language is mainly used in The Netherlands and Flanders (Belgium). I chose nl_NL as the locale identifier, but both are really the same. I used polite and formal language (e.g. "u" instead of "je", which both means "you", but "je" usually is used informally).

Each line is translated from the English version. some lines have been reworded to comply with our language, but they carry the exact same meaning.

In case of issues, please feel free to contact me in order to look into it.

(If you wish, I can also send a slightly customized theme (cyan text on the top Menu Bar, based on the default theme, so really only a text colour change) but I hope this contribution is useful and helps a bit.

Kind regards