Added command line args for controlling shell mode
/noshell - force disable shell mode
/shell - force enable shell mode
This can be used to force-disable shell mode if Cairo is starting as a logon scheduled task. Otherwise, this could race with Explorer and we could make bad decisions.
Previously this wasn't an issue because we would only enable shell mode if we were set as such in the registry.