A little change in implementation to allow the changing of AllowsTransparency via theming.
NOTE: this setting has to be set prior to the window creation and cannot change afterwards. Due to this, I have added the resource consumption to the constructors of MenuBar and TaskBar for now. This means that any theme change that uses these new settings require the effected windows to be recreated.